call

+++ crates/everything/src/base.rs
@@ -25,7 +25,7 @@ fn stmt_to_prop(subject: Object, tag: Object, value: Object) -> Property {
     }
 }
 
-fn common_unique_constraint_expression(tag: Object, parameter_depth: u64) -> Object {
+fn common_unique_constraint_expression(tag: Object, parameter_depth: usize) -> Object {
     Structure::new_node_equal([
         Object::natural_number(1),
         Structure::new_node_count(
@@ -61,11 +61,11 @@ fn common_unique_constraint_expression(tag: Object, parameter_depth: u64) -> Obj
 /// ```plain
 /// ... |-> count query {(@4, $parameter_at_depth), (@5, tag)} == 1
 /// ```
-fn unique_constraint_for(tag: Object, parameter_depth: u64) -> Object {
+fn unique_constraint_for(tag: Object, parameter_depth: usize) -> Object {
     Structure::new_computed(common_unique_constraint_expression(tag, parameter_depth)).into()
 }
 
-fn more_than_1_constraint_for(tag: Object, parameter_depth: u64) -> Object {
+fn more_than_1_constraint_for(tag: Object, parameter_depth: usize) -> Object {
     Structure::new_computed(
         Structure::new_node_not(common_unique_constraint_expression(tag, parameter_depth)).into(),
     )
